+++ packages/astro/src/@types/astro.ts
@@ -281,7 +281,7 @@ export interface AstroGlobal<
 		 * <Fragment set:html={html} />
 		 * ```
 		 *
-		 * A second parameters can be used to pass arguments to a slotted callback
+		 * A second parameter can be used to pass arguments to a slotted callback
 		 *
 		 * Example usage:
 		 * ```astro
@@ -2839,6 +2839,7 @@ export interface PreviewServerParams {
 	port: number;
 	base: string;
 	logger: AstroIntegrationLogger;
+	headers?: OutgoingHttpHeaders;
 }
 
 export type CreatePreviewServer = (