+++ packages/db/config-augment.d.ts
@@ -1,0 +1,4 @@
+declare namespace Config {
+	type DBUserConfig = import('./dist/core/types.js').DBUserConfig;
+	export interface Database extends DBUserConfig {}
+}