+++ packages/db/src/core/consts.ts
@@ -1,0 +1,14 @@
+import { readFileSync } from 'node:fs';
+
+export const PACKAGE_NAME = JSON.parse(
+	readFileSync(new URL('../../package.json', import.meta.url), 'utf8')
+).name;
+
+export const RUNTIME_IMPORT = JSON.stringify(`${PACKAGE_NAME}/runtime`);
+export const RUNTIME_DRIZZLE_IMPORT = JSON.stringify(`${PACKAGE_NAME}/runtime/drizzle`);
+
+export const DB_TYPES_FILE = 'db-types.d.ts';
+
+export const VIRTUAL_MODULE_ID = 'astro:db';
+
+export const DB_PATH = '.astro/content.db';